refactor: split domain into model and service
refactor: split application into service
