+++ src/taxpasta/application/__init__.py
@@ -13,9 +13,5 @@
 # limitations under the License.
 
 
-from .profile_reader import ProfileReader
-from .profile_standardisation_service import ProfileStandardisationService
 from .sample_merging_application import SampleMergingApplication
-from ._types import BufferOrFilepath, BinaryBufferOrFilepath, Filepath
-from .table_reader import TableReader
-from .table_writer import TableWriter
+from .consensus_application import ConsensusApplication
