+++ src/taxpasta/application/service/table_writer.py
