+++ src/taxpasta/domain/__init__.py
@@ -11,12 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-
-from .standard_profile import StandardProfile
-from .sample import Sample
-from .tidy_observation_table import TidyObservationTable
-from .observation_matrix import ObservationMatrix
-from .taxonomy import Taxonomy
-from .sample_merging_service import SampleMergingService
-from .consensus_builder import ConsensusBuilder
