+++ src/taxpasta/infrastructure/application/bracken/bracken_profile_reader.py
@@ -19,7 +19,7 @@
 import pandas as pd
 from pandera.typing import DataFrame
 
-from taxpasta.application import BufferOrFilepath, ProfileReader
+from taxpasta.application.service import BufferOrFilepath, ProfileReader
 
 from .bracken_profile import BrackenProfile
