+++ src/taxpasta/infrastructure/application/centrifuge/centrifuge_profile_standardisation_service.py
@@ -21,8 +21,8 @@ import logging
 import pandera as pa
 from pandera.typing import DataFrame
 
-from taxpasta.application import ProfileStandardisationService
-from taxpasta.domain import StandardProfile
+from taxpasta.application.service import ProfileStandardisationService
+from taxpasta.domain.model import StandardProfile
 
 from .centrifuge_profile import CentrifugeProfile
