+++ src/taxpasta/infrastructure/application/kaiju/kaiju_profile_reader.py
@@ -19,7 +19,7 @@
 import pandas as pd
 from pandera.typing import DataFrame
 
-from taxpasta.application import BufferOrFilepath, ProfileReader
+from taxpasta.application.service import BufferOrFilepath, ProfileReader
 
 from .kaiju_profile import KaijuProfile
