+++ src/taxpasta/infrastructure/application/kaiju/kaiju_profile_standardisation_service.py
@@ -19,8 +19,8 @@
 import pandera as pa
 from pandera.typing import DataFrame
 
-from taxpasta.application import ProfileStandardisationService
-from taxpasta.domain import StandardProfile
+from taxpasta.application.service import ProfileStandardisationService
+from taxpasta.domain.model import StandardProfile
 
 from .kaiju_profile import KaijuProfile
