+++ src/taxpasta/infrastructure/application/kraken2/kraken2_profile_reader.py
@@ -19,7 +19,7 @@
 import pandas as pd
 from pandera.typing import DataFrame
 
-from taxpasta.application import BufferOrFilepath, ProfileReader
+from taxpasta.application.service import BufferOrFilepath, ProfileReader
 
 from .kraken2_profile import Kraken2Profile
