+++ src/taxpasta/infrastructure/application/metaphlan/metaphlan_profile_reader.py
@@ -18,7 +18,7 @@
 import pandas as pd
 from pandera.typing import DataFrame
 
-from taxpasta.application import BufferOrFilepath, ProfileReader
+from taxpasta.application.service import BufferOrFilepath, ProfileReader
 
 from .metaphlan_profile import MetaphlanProfile
