+++ src/taxpasta/infrastructure/application/metaphlan/metaphlan_profile_standardisation_service.py
@@ -14,12 +14,14 @@
 
 
 """Provide a standardisation service for metaphlan profiles."""
+
+
 import pandas as pd
 import pandera as pa
 from pandera.typing import DataFrame
 
-from taxpasta.application import ProfileStandardisationService
-from taxpasta.domain import StandardProfile
+from taxpasta.application.service import ProfileStandardisationService
+from taxpasta.domain.model import StandardProfile
 
 from .metaphlan_profile import MetaphlanProfile
