+++ src/taxpasta/infrastructure/application/table_reader/xlsx_table_reader.py
@@ -18,7 +18,7 @@
 
 import pandas as pd
 
-from taxpasta.application import BinaryBufferOrFilepath, TableReader
+from taxpasta.application.service import BinaryBufferOrFilepath, TableReader
 
 
 class XLSXTableReader(TableReader):
